+++ src/xgboost_distribution/distributions/poisson.py
@@ -10,6 +10,8 @@ from xgboost_distribution.distributions.utils import check_is_integer, check_is_
 class Poisson(BaseDistribution):
     """Poisson distribution with log score
 
+    Definition:
+
         f(k) = e^(-mu) mu^k / k!
 
     We reparameterize mu -> log(mu) = a to ensure mu >= 0. Gradient: