+++ src/xgboost_distribution/distributions/utils.py
@@ -10,3 +10,8 @@ def check_is_integer(x):
 def check_is_positive(x):
     if not all(x >= 0):
         raise ValueError("All values of target must be positive!")
+
+
+def check_is_gt_zero(x):
+    if not all(x > 0):
+        raise ValueError("All values of target must be > 0!")