+++ tests/distributions/test_base.py
@@ -4,6 +4,7 @@ import pytest
 
 from xgboost_distribution.distributions import (
     Normal,
+    format_distribution_name,
     get_distribution,
     get_distribution_doc,
 )
@@ -24,3 +25,12 @@ def test_get_distribution_raises():
 def test_get_distribution_doc():
     distribution_doc = get_distribution_doc()
     assert isinstance(distribution_doc, str)
+
+
+@pytest.mark.parametrize(
+    "name, expected_name",
+    [("LogNormal", "log-normal"), ("NegativeBinomial", "negative-binomial")],
+)
+def test_format_distribution_name(name, expected_name):
+    formatted_name = format_distribution_name(name)
+    assert formatted_name == expected_name